Switch to uint64_t in Problem301.cpp
Using a normal int was causing an integer overflow.

+++ src/Problem301.cpp
@@ -45,8 +45,10 @@
  */
 
 #include <iostream>  // std::cout, std::endl
+#include <inttypes.h>  // uint64_t
 
-bool IsLoss(const int& a, const int& b, const int& c) {
+
+bool IsLoss(const uint64_t& a, const uint64_t& b, const uint64_t& c) {
     // If the xor of the positions is 0, then the next player will lose with
     // perfect play.
     if ((a ^ b ^ c) == 0) {
@@ -57,11 +59,11 @@ bool IsLoss(const int& a, const int& b, const int& c) {
 }
 
 int main() {
-    const int MAX = 1073741824;
-    int counter = 0;
+    const uint64_t MAX = 1073741824;
+    uint64_t counter = 0;
 
     // Test all possible games
-    for (int i = 1; i <= MAX; ++i) {
+    for (uint64_t i = 1; i <= MAX; ++i) {
         if (IsLoss(i, 2 * i, 3 * i)) {
             counter++;
         }
